Move test to SupportModels directory
